ai-engine/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
pymupdf==1.23.8
python-docx==1.1.0
transformers==4.37.2
torch==2.1.2
sentencepiece==0.1.99
python-multipart==0.0.6
pydantic==2.5.3
python-dotenv==1.0.0
faiss-cpu==1.7.4
sentence-transformers==2.3.1
numpy==1.26.3
